+++ pytripgui/main_window_qt_vc/main_window_view.py
@@ -5,7 +5,6 @@ from PyQt5.QtWidgets import QFileDialog
 from pytripgui.view.qt_gui import UiMainWindow
 from pytripgui.config_vc import ConfigQtView
 from pytripgui.kernel_vc import KernelQtView
-from pytripgui.treewidget_vc.treewidget_view import TreeWidgetView
 from pytripgui.viewcanvas_vc.viewcanvas_view import ViewCanvasView
 
 logger = logging.getLogger(__name__)
