+++ pytripgui/tree_vc/TreeItems.py
@@ -9,14 +9,42 @@ logger = logging.getLogger(__name__)
 
 
 class PatientList(NodeMixin):
-    def __init__(self):
+    """
+    Root item in TreeModel
+    """
+    def __init__(self, name="PatientList"):
         super().__init__()
+        self.name = name
         self.parent = None
-        self.data = list()
 
-    def _post_attach_children(self, children):
-        for child in children:
-            self.data.append(child.data)
+    def __repr__(self):
+        return self.name
+
+    def add_child(self, child):
+        if isinstance(child, PatientItem):
+            self.children += tuple([child])
+        else:
+            raise Exception("PatientItem can only be added to PatientList")
+
+    def has_index(self, p_int):
+        return p_int < len(self.children)
+
+    # For qt TreeView
+    def has_children(self):
+        return len(self.children) > 0
+
+    def row_count(self):
+        return len(self.children)
+
+    def index(self, p_int, p_int_1, obj):
+        from PyQt5.QtCore import QModelIndex
+
+        if not self.has_index(p_int) or \
+                p_int_1 != 0:   # only one column is supported
+            return QModelIndex()
+
+        index = obj.createIndex(p_int, p_int_1, self.children[p_int])
+        return index
 
 
 class PatientItem(NodeMixin):
@@ -34,10 +62,12 @@ class PatientItem(NodeMixin):
         if parent and not isinstance(parent, PatientList):
             raise Exception("FieldItems can only be added to PlanItem")
 
-    def _post_attach_children(self, children):
-        print("Adding plan to PatientItem")
-        for child in children:
-            self.data.plans.append(child.data)
+    # For qt TreeView
+    def has_children(self):
+        return len(self.children) > 0
+
+    def __repr__(self):
+        return self.data.name
 
 
 class PlanItem(NodeMixin):
@@ -60,7 +90,6 @@ class FieldItem(NodeMixin):
     def __init__(self, data=Field(), parent=None, children=None):
         super().__init__()
         self.parent = parent
-        self.name = "sad"
         if children:
             self.children = children
 
@@ -78,7 +107,8 @@ patient_list = PatientList()
 udo = PatientItem()
 udo1 = PatientItem()
 
-patient_list.children += tuple([udo, udo1])
+patient_list.add_child(udo)
+patient_list.add_child(udo1)
 patient_list.children += tuple([PatientItem()])
 
 udo.children += tuple([PlanItem()])
